Refactor to two modules and filter the set of returned data points from the Service to a reasonable number
