+++ application/Dockerfile
@@ -1,0 +1,5 @@
+FROM jetty:9.4.9
+
+ENV STOCKS_MYSQL_HOST db
+
+COPY target/application-1.0-SNAPSHOT.war /var/lib/jetty/webapps/ROOT.war
