+++ application/src/main/java/com/vaadin/demo/stockdata/ui/MoneyFormatter.java
