+++ application/src/main/java/com/vaadin/demo/stockdata/ui/XAxisExtremesEvent.java
