+++ application/src/main/webapp/frontend/src/sparkline-chart.html
