+++ application/src/main/webapp/frontend/src/speedment-chart-theme.html
