+++ application/src/main/webapp/frontend/styles/shared-styles.html
