+++ application/src/main/webapp/frontend/styles/stock-details.css
