+++ application/src/main/webapp/frontend/styles/stock-list.css
