+++ data/src/main/resources/nasdaq_tickers.csv
